moved cmd/otel/components -> internal/components
