+++ internal/components/components.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package components
 
 import (
 	"go.opentelemetry.io/collector/component"
@@ -89,7 +89,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/zookeeperreceiver"
 )
 
-func components() (component.Factories, error) {
+func Components() (component.Factories, error) {
 	factories, err := defaultcomponents.Components()
 	if err != nil {
 		return component.Factories{}, err
@@ -143,7 +143,7 @@ func components() (component.Factories, error) {
 		udplogreceiver.NewFactory(),
 	}
 
-	receivers = append(receivers, extraReceivers()...)
+	receivers = append(receivers, []component.ReceiverFactory{}...)
 
 	for _, rcv := range factories.Receivers {
 		receivers = append(receivers, rcv)
